refactor(server): split large server.go into organized server package
Split the monolithic server.go (1116 lines) into a well-organized
server package with focused, maintainable files:

- server/server.go (112 lines): Core Server struct and constructor
- server/config.go (170 lines): Configuration and secure defaults
- server/flows.go (555 lines): OAuth authorization flows
- server/validation.go (268 lines): PKCE and redirect URI validation
- server/client.go (100 lines): Client registration and management
- server/doc.go (35 lines): Package documentation

The root server.go (30 lines) now serves as a thin compatibility
layer, maintaining backward compatibility for existing code.

Benefits:
- Improved code organization and maintainability
- Easier to navigate and understand specific functionality
- Better separation of concerns (config, flows, validation, clients)
- All files are well under 500 lines (largest is 555 lines)
- No breaking changes - full backward compatibility maintained

Updated handler.go to use exported Server fields (Config, Auditor,
RateLimiter, etc.) as they are now accessed from outside the package.
